EZP-27743: Rename Criterion abstract class to Matcher
