+++ doc/bc/changes-7.2.md
@@ -1,0 +1,13 @@
+# Backwards compatibility changes
+
+Changes affecting version compatibility with former or future versions.
+
+## Changes
+
+## Deprecations
+
+- The `eZ\Publish\API\Repository\Values\Content\Query\Criterion` class is deprecated and will be removed in 8.0.
+
+  Use `eZ\Publish\API\Repository\Values\Content\Query\Criterion\Matcher\Matcher` instead.
+
+## Removed features
